chore: Move StarshipConfig to a separate file
